Update 21-0845 e2e-spec
