--- src/applications/simple-forms/21-0845/tests/0845-auth-disclose.cypress.spec.js
@@ -1,43 +1,0 @@
-import path from 'path';
-
-import testForm from 'platform/testing/e2e/cypress/support/form-tester';
-import { createTestConfig } from 'platform/testing/e2e/cypress/support/form-tester/utilities';
-
-import formConfig from '../config/form';
-import manifest from '../manifest.json';
-
-const testConfig = createTestConfig(
-  {
-    dataPrefix: 'data',
-
-    dataDir: path.join(__dirname, 'data'),
-
-    // Rename and modify the test data as needed.
-    dataSets: ['testdata'],
-
-    pageHooks: {
-      introduction: ({ afterHook }) => {
-        afterHook(() => {
-          cy.findAllByText(/start/i, { selector: 'button' })
-            .last()
-            .click();
-        });
-      },
-    },
-
-    setupPerTest: () => {
-      // Log in if the form requires an authenticated session.
-      // cy.login();
-
-      cy.route('POST', formConfig.submitUrl, { status: 200 });
-    },
-
-    // Skip tests in CI until the form is released.
-    // Remove this setting when the form has a content page in production.
-    skip: Cypress.env('CI'),
-  },
-  manifest,
-  formConfig,
-);
-
-testForm(testConfig);
